Bringing the OSMAnd fork on to the same version as the original repository

+++ OsmAnd/src/net/osmand/plus/myplaces/TrackSegmentFragment.java
@@ -11,6 +11,7 @@ import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
+import android.support.v4.content.ContextCompat;
 import android.support.v4.view.MenuItemCompat;
 import android.support.v4.view.PagerAdapter;
 import android.support.v4.view.ViewCompat;
@@ -38,24 +39,24 @@ import com.github.mikephil.charting.listener.OnChartGestureListener;
 import com.github.mikephil.charting.listener.OnChartValueSelectedListener;
 
 import net.osmand.AndroidUtils;
-import net.osmand.data.LatLon;
-import net.osmand.data.PointDescription;
-import net.osmand.plus.GPXDatabase.GpxDataItem;
 import net.osmand.GPXUtilities;
 import net.osmand.GPXUtilities.GPXFile;
 import net.osmand.GPXUtilities.GPXTrackAnalysis;
 import net.osmand.GPXUtilities.Track;
 import net.osmand.GPXUtilities.TrkSegment;
 import net.osmand.GPXUtilities.WptPt;
+import net.osmand.data.LatLon;
+import net.osmand.data.PointDescription;
+import net.osmand.plus.GPXDatabase.GpxDataItem;
 import net.osmand.plus.GpxSelectionHelper.GpxDisplayGroup;
 import net.osmand.plus.GpxSelectionHelper.GpxDisplayItem;
 import net.osmand.plus.GpxSelectionHelper.GpxDisplayItemType;
 import net.osmand.plus.GpxSelectionHelper.SelectedGpxFile;
-import net.osmand.plus.UiUtilities;
 import net.osmand.plus.OsmAndFormatter;
 import net.osmand.plus.OsmandApplication;
 import net.osmand.plus.OsmandSettings;
 import net.osmand.plus.R;
+import net.osmand.plus.UiUtilities;
 import net.osmand.plus.activities.MapActivity;
 import net.osmand.plus.activities.TrackActivity;
 import net.osmand.plus.base.OsmAndListFragment;
@@ -332,13 +333,15 @@ public class TrackSegmentFragment extends OsmAndListFragment implements TrackBit
 				boolean light = app.getSettings().isLightContent();
 				tabLayout = (PagerSlidingTabStrip) row.findViewById(R.id.sliding_tabs);
 				tabLayout.setTabBackground(R.color.color_transparent);
-				tabLayout.setIndicatorColorResource(light ? R.color.color_dialog_buttons_light : R.color.color_dialog_buttons_dark);
-				tabLayout.setIndicatorBgColorResource(light ? R.color.dashboard_divider_light : R.color.dashboard_divider_dark);
+				tabLayout.setIndicatorColorResource(light ? R.color.active_color_primary_light : R.color.active_color_primary_dark);
+				tabLayout.setIndicatorBgColorResource(light ? R.color.divider_color_light : R.color.divider_color_dark);
 				tabLayout.setIndicatorHeight(AndroidUtils.dpToPx(app, 1f));
-				tabLayout.setTextColor(tabLayout.getIndicatorColor());
+				if (light) {
+					tabLayout.setTextColor(tabLayout.getIndicatorColor());
+					tabLayout.setTabInactiveTextColor(ContextCompat.getColor(row.getContext(), R.color.text_color_secondary_light));
+				}
 				tabLayout.setTextSize(AndroidUtils.spToPx(app, 12f));
 				tabLayout.setShouldExpand(true);
-				tabLayout.setTabSelectionType(PagerSlidingTabStrip.TabSelectionType.SOLID_COLOR);
 				pager = (WrapContentHeightViewPager) row.findViewById(R.id.pager);
 				pager.setSwipeable(false);
 				pager.setOffscreenPageLimit(2);
@@ -376,6 +379,7 @@ public class TrackSegmentFragment extends OsmAndListFragment implements TrackBit
 		private Map<GPXTabItemType, List<ILineDataSet>> dataSetsMap = new HashMap<>();
 		private TrkSegment segment;
 		private float listViewYPos;
+		private WptPt selectedWpt;
 
 		GPXItemPagerAdapter(PagerSlidingTabStrip tabs, GpxDisplayItem gpxItem) {
 			super();
@@ -550,7 +554,6 @@ public class TrackSegmentFragment extends OsmAndListFragment implements TrackBit
 		@Override
 		public Object instantiateItem(@NonNull ViewGroup container, int position) {
 
-			chartClicked = false;
 			GPXTabItemType tabType = tabTypes[position];
 			final View view;
 			LayoutInflater inflater = LayoutInflater.from(container.getContext());
@@ -572,109 +575,118 @@ public class TrackSegmentFragment extends OsmAndListFragment implements TrackBit
 			if (gpxFile != null && gpxItem != null) {
 				GPXTrackAnalysis analysis = gpxItem.analysis;
 				final LineChart chart = (LineChart) view.findViewById(R.id.chart);
-				chart.setHighlightPerDragEnabled(false);
+				chart.setHighlightPerDragEnabled(chartClicked);
 				chart.setOnClickListener(new View.OnClickListener() {
 					@SuppressLint("ClickableViewAccessibility")
 					@Override
 					public void onClick(View view) {
 						if (!chartClicked) {
 							chartClicked = true;
-							chart.setHighlightPerDragEnabled(true);
-							chart.setOnTouchListener(new View.OnTouchListener() {
-								@Override
-								public boolean onTouch(View v, MotionEvent event) {
-									getListView().requestDisallowInterceptTouchEvent(true);
-									switch (event.getAction()) {
-										case android.view.MotionEvent.ACTION_DOWN:
-											listViewYPos = event.getRawY();
-											break;
-										case android.view.MotionEvent.ACTION_MOVE:
-											scrollBy(Math.round(listViewYPos - event.getRawY()));
-											listViewYPos = event.getRawY();
-											break;
-									}
-									return false;
-								}
-							});
-							chart.setOnChartValueSelectedListener(new OnChartValueSelectedListener() {
-								@Override
-								public void onValueSelected(Entry e, Highlight h) {
-									WptPt wpt = getPoint(chart, h.getX());
-									if (wpt != null && fragmentAdapter != null) {
-										fragmentAdapter.updateSelectedPoint(wpt.lat, wpt.lon);
-									}
-								}
+							if (selectedWpt != null && fragmentAdapter != null) {
+								fragmentAdapter.updateSelectedPoint(selectedWpt.lat, selectedWpt.lon);
+							}
+						}
+					}
+				});
+				chart.setOnTouchListener(new View.OnTouchListener() {
+					@Override
+					public boolean onTouch(View v, MotionEvent event) {
+						if (chartClicked) {
+							getListView().requestDisallowInterceptTouchEvent(true);
+							if (!chart.isHighlightPerDragEnabled()) {
+								chart.setHighlightPerDragEnabled(true);
+							}
+							switch (event.getAction()) {
+								case android.view.MotionEvent.ACTION_DOWN:
+									listViewYPos = event.getRawY();
+									break;
+								case android.view.MotionEvent.ACTION_MOVE:
+									scrollBy(Math.round(listViewYPos - event.getRawY()));
+									listViewYPos = event.getRawY();
+									break;
+							}
+						}
+						return false;
+					}
+				});
+				chart.setOnChartValueSelectedListener(new OnChartValueSelectedListener() {
+					@Override
+					public void onValueSelected(Entry e, Highlight h) {
+						WptPt wpt = getPoint(chart, h.getX());
+						selectedWpt = wpt;
+						if (chartClicked && wpt != null && fragmentAdapter != null) {
+							fragmentAdapter.updateSelectedPoint(wpt.lat, wpt.lon);
+						}
+					}
 
-								@Override
-								public void onNothingSelected() {
+					@Override
+					public void onNothingSelected() {
 
+					}
+				});
+				chart.setOnChartGestureListener(new OnChartGestureListener() {
+
+					float highlightDrawX = -1;
+
+					@Override
+					public void onChartGestureStart(MotionEvent me, ChartGesture lastPerformedGesture) {
+						if (chart.getHighlighted() != null && chart.getHighlighted().length > 0) {
+							highlightDrawX = chart.getHighlighted()[0].getDrawX();
+						} else {
+							highlightDrawX = -1;
+						}
+					}
+
+					@Override
+					public void onChartGestureEnd(MotionEvent me, ChartGesture lastPerformedGesture) {
+						gpxItem.chartMatrix = new Matrix(chart.getViewPortHandler().getMatrixTouch());
+						Highlight[] highlights = chart.getHighlighted();
+						if (highlights != null && highlights.length > 0) {
+							gpxItem.chartHighlightPos = highlights[0].getX();
+						} else {
+							gpxItem.chartHighlightPos = -1;
+						}
+						if (chartClicked) {
+							for (int i = 0; i < getCount(); i++) {
+								View v = getViewAtPosition(i);
+								if (v != view) {
+									updateChart(i);
 								}
-							});
-							final View finalView = view;
-							chart.setOnChartGestureListener(new OnChartGestureListener() {
+							}
+						}
+					}
 
-								float highlightDrawX = -1;
+					@Override
+					public void onChartLongPressed(MotionEvent me) {
+					}
 
-								@Override
-								public void onChartGestureStart(MotionEvent me, ChartGesture lastPerformedGesture) {
-									if (chart.getHighlighted() != null && chart.getHighlighted().length > 0) {
-										highlightDrawX = chart.getHighlighted()[0].getDrawX();
-									} else {
-										highlightDrawX = -1;
-									}
+					@Override
+					public void onChartDoubleTapped(MotionEvent me) {
+					}
+
+					@Override
+					public void onChartSingleTapped(MotionEvent me) {
+					}
+
+					@Override
+					public void onChartFling(MotionEvent me1, MotionEvent me2, float velocityX, float velocityY) {
+					}
+
+					@Override
+					public void onChartScale(MotionEvent me, float scaleX, float scaleY) {
+					}
+
+					@Override
+					public void onChartTranslate(MotionEvent me, float dX, float dY) {
+						if (chartClicked && highlightDrawX != -1) {
+							Highlight h = chart.getHighlightByTouchPoint(highlightDrawX, 0f);
+							if (h != null) {
+								chart.highlightValue(h);
+								WptPt wpt = getPoint(chart, h.getX());
+								if (wpt != null && fragmentAdapter != null) {
+									fragmentAdapter.updateSelectedPoint(wpt.lat, wpt.lon);
 								}
-
-								@Override
-								public void onChartGestureEnd(MotionEvent me, ChartGesture lastPerformedGesture) {
-									gpxItem.chartMatrix = new Matrix(chart.getViewPortHandler().getMatrixTouch());
-									Highlight[] highlights = chart.getHighlighted();
-									if (highlights != null && highlights.length > 0) {
-										gpxItem.chartHighlightPos = highlights[0].getX();
-									} else {
-										gpxItem.chartHighlightPos = -1;
-									}
-									for (int i = 0; i < getCount(); i++) {
-										View v = getViewAtPosition(i);
-										if (v != finalView) {
-											updateChart(i);
-										}
-									}
-								}
-
-								@Override
-								public void onChartLongPressed(MotionEvent me) {
-								}
-
-								@Override
-								public void onChartDoubleTapped(MotionEvent me) {
-								}
-
-								@Override
-								public void onChartSingleTapped(MotionEvent me) {
-								}
-
-								@Override
-								public void onChartFling(MotionEvent me1, MotionEvent me2, float velocityX, float velocityY) {
-								}
-
-								@Override
-								public void onChartScale(MotionEvent me, float scaleX, float scaleY) {
-								}
-
-								@Override
-								public void onChartTranslate(MotionEvent me, float dX, float dY) {
-									if (highlightDrawX != -1) {
-										Highlight h = chart.getHighlightByTouchPoint(highlightDrawX, 0f);
-										if (h != null) {
-											chart.highlightValue(h);
-											WptPt wpt = getPoint(chart, h.getX());
-											if (wpt != null && fragmentAdapter != null) {
-												fragmentAdapter.updateSelectedPoint(wpt.lat, wpt.lon);
-											}
-										}
-									}
-								}
-							});
+							}
 						}
 					}
 				});
@@ -758,37 +770,37 @@ public class TrackSegmentFragment extends OsmAndListFragment implements TrackBit
 										generalPopupMenu.setOnMenuItemClickListener(new IconPopupMenu.OnMenuItemClickListener() {
 											@Override
 											public boolean onMenuItemClick(MenuItem item) {
-												switch (item.getItemId()) {
-													case R.id.action_edit:
-														TrkSegment segment = getTrkSegment();
-														if (segment != null && fragmentAdapter != null) {
-															fragmentAdapter.addNewGpxData(NewGpxData.ActionType.EDIT_SEGMENT, segment);
-														}
-														return true;
-													case R.id.action_delete:
-														TrackActivity activity = getTrackActivity();
-														if (activity != null) {
-															AlertDialog.Builder builder = new AlertDialog.Builder(activity);
-															builder.setMessage(R.string.recording_delete_confirm);
-															builder.setPositiveButton(R.string.shared_string_yes, new DialogInterface.OnClickListener() {
-																@Override
-																public void onClick(DialogInterface dialog, int which) {
-																	TrackActivity trackActivity = getTrackActivity();
-																	if (trackActivity != null && deleteSegment()) {
-																		GPXFile gpx = getGpx();
-																		if (gpx != null && fragmentAdapter != null) {
-																			boolean showOnMap = fragmentAdapter.isShowOnMap();
-																			SelectedGpxFile sf = app.getSelectedGpxHelper().selectGpxFile(gpx, showOnMap, false);
-																			new SaveGpxAsyncTask(trackActivity, TrackSegmentFragment.this, gpx, showOnMap ? sf : null)
-																					.executeOnExecutor(AsyncTask.THREAD_POOL_EXECUTOR);
-																		}
+												int i = item.getItemId();
+												if (i == R.id.action_edit) {
+													TrkSegment segment = getTrkSegment();
+													if (segment != null && fragmentAdapter != null) {
+														fragmentAdapter.addNewGpxData(NewGpxData.ActionType.EDIT_SEGMENT, segment);
+													}
+													return true;
+												} else if (i == R.id.action_delete) {
+													TrackActivity activity = getTrackActivity();
+													if (activity != null) {
+														AlertDialog.Builder builder = new AlertDialog.Builder(activity);
+														builder.setMessage(R.string.recording_delete_confirm);
+														builder.setPositiveButton(R.string.shared_string_yes, new DialogInterface.OnClickListener() {
+															@Override
+															public void onClick(DialogInterface dialog, int which) {
+																TrackActivity trackActivity = getTrackActivity();
+																if (trackActivity != null && deleteSegment()) {
+																	GPXFile gpx = getGpx();
+																	if (gpx != null && fragmentAdapter != null) {
+																		boolean showOnMap = fragmentAdapter.isShowOnMap();
+																		SelectedGpxFile sf = app.getSelectedGpxHelper().selectGpxFile(gpx, showOnMap, false);
+																		new SaveGpxAsyncTask(trackActivity, TrackSegmentFragment.this, gpx, showOnMap ? sf : null)
+																				.executeOnExecutor(AsyncTask.THREAD_POOL_EXECUTOR);
 																	}
 																}
-															});
-															builder.setNegativeButton(R.string.shared_string_cancel, null);
-															builder.show();
-														}
-														return true;
+															}
+														});
+														builder.setNegativeButton(R.string.shared_string_cancel, null);
+														builder.show();
+													}
+													return true;
 												}
 												return false;
 											}
@@ -870,25 +882,25 @@ public class TrackSegmentFragment extends OsmAndListFragment implements TrackBit
 										altitudePopupMenu.setOnMenuItemClickListener(new IconPopupMenu.OnMenuItemClickListener() {
 											@Override
 											public boolean onMenuItemClick(MenuItem item) {
-												switch (item.getItemId()) {
-													case R.id.action_edit:
-														TrkSegment segment = getTrkSegment();
-														if (segment != null && fragmentAdapter != null) {
-															fragmentAdapter.addNewGpxData(NewGpxData.ActionType.EDIT_SEGMENT, segment);
+												int i = item.getItemId();
+												if (i == R.id.action_edit) {
+													TrkSegment segment = getTrkSegment();
+													if (segment != null && fragmentAdapter != null) {
+														fragmentAdapter.addNewGpxData(NewGpxData.ActionType.EDIT_SEGMENT, segment);
+													}
+													return true;
+												} else if (i == R.id.action_delete) {
+													TrackActivity trackActivity = getTrackActivity();
+													if (trackActivity != null && deleteSegment()) {
+														GPXFile gpx = getGpx();
+														if (gpx != null && fragmentAdapter != null) {
+															boolean showOnMap = fragmentAdapter.isShowOnMap();
+															SelectedGpxFile sf = app.getSelectedGpxHelper().selectGpxFile(gpx, showOnMap, false);
+															new SaveGpxAsyncTask(trackActivity, TrackSegmentFragment.this, gpx, showOnMap ? sf : null)
+																	.executeOnExecutor(AsyncTask.THREAD_POOL_EXECUTOR);
 														}
-														return true;
-													case R.id.action_delete:
-														TrackActivity trackActivity = getTrackActivity();
-														if (trackActivity != null && deleteSegment()) {
-															GPXFile gpx = getGpx();
-															if (gpx != null && fragmentAdapter != null) {
-																boolean showOnMap = fragmentAdapter.isShowOnMap();
-																SelectedGpxFile sf = app.getSelectedGpxHelper().selectGpxFile(gpx, showOnMap, false);
-																new SaveGpxAsyncTask(trackActivity, TrackSegmentFragment.this, gpx, showOnMap ? sf : null)
-																		.executeOnExecutor(AsyncTask.THREAD_POOL_EXECUTOR);
-															}
-														}
-														return true;
+													}
+													return true;
 												}
 												return false;
 											}
@@ -969,25 +981,25 @@ public class TrackSegmentFragment extends OsmAndListFragment implements TrackBit
 										popupMenu.setOnMenuItemClickListener(new IconPopupMenu.OnMenuItemClickListener() {
 											@Override
 											public boolean onMenuItemClick(MenuItem item) {
-												switch (item.getItemId()) {
-													case R.id.action_edit:
-														TrkSegment segment = getTrkSegment();
-														if (segment != null && fragmentAdapter != null) {
-															fragmentAdapter.addNewGpxData(NewGpxData.ActionType.EDIT_SEGMENT, segment);
+												int i = item.getItemId();
+												if (i == R.id.action_edit) {
+													TrkSegment segment = getTrkSegment();
+													if (segment != null && fragmentAdapter != null) {
+														fragmentAdapter.addNewGpxData(NewGpxData.ActionType.EDIT_SEGMENT, segment);
+													}
+													return true;
+												} else if (i == R.id.action_delete) {
+													TrackActivity trackActivity = getTrackActivity();
+													if (trackActivity != null && deleteSegment()) {
+														GPXFile gpx = getGpx();
+														if (gpx != null && fragmentAdapter != null) {
+															boolean showOnMap = fragmentAdapter.isShowOnMap();
+															SelectedGpxFile sf = app.getSelectedGpxHelper().selectGpxFile(gpx, showOnMap, false);
+															new SaveGpxAsyncTask(trackActivity, TrackSegmentFragment.this, gpx, showOnMap ? sf : null)
+																	.executeOnExecutor(AsyncTask.THREAD_POOL_EXECUTOR);
 														}
-														return true;
-													case R.id.action_delete:
-														TrackActivity trackActivity = getTrackActivity();
-														if (trackActivity != null && deleteSegment()) {
-															GPXFile gpx = getGpx();
-															if (gpx != null && fragmentAdapter != null) {
-																boolean showOnMap = fragmentAdapter.isShowOnMap();
-																SelectedGpxFile sf = app.getSelectedGpxHelper().selectGpxFile(gpx, showOnMap, false);
-																new SaveGpxAsyncTask(trackActivity, TrackSegmentFragment.this, gpx, showOnMap ? sf : null)
-																		.executeOnExecutor(AsyncTask.THREAD_POOL_EXECUTOR);
-															}
-														}
-														return true;
+													}
+													return true;
 												}
 												return false;
 											}
@@ -1113,9 +1125,11 @@ public class TrackSegmentFragment extends OsmAndListFragment implements TrackBit
 
 		private TrkSegment getTrkSegment() {
 			for (Track t : gpxItem.group.getGpx().tracks) {
-				for (TrkSegment s : t.segments) {
-					if (s.points.size() > 0 && s.points.get(0).equals(gpxItem.analysis.locationStart)) {
-						return s;
+				if (!t.generalTrack) {
+					for (TrkSegment s : t.segments) {
+						if (s.points.size() > 0 && s.points.get(0).equals(gpxItem.analysis.locationStart)) {
+							return s;
+						}
 					}
 				}
 			}
